requirements.txt
numpy
pillow
pytesseract
requests
selenium